+++ packages/alova/CHANGELOG.md
@@ -1,5 +1,29 @@
 # alova
 
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- [#432](https://github.com/alovajs/alova/pull/432) [`586103f`](https://github.com/alovajs/alova/commit/586103fefdb3ddc1ae8eab40f6c92ebab4f495a8) Thanks [@JOU-amjs](https://github.com/JOU-amjs)! - fix that states updating not work with `updateStateEffect` in react
+
+- Updated dependencies [[`586103f`](https://github.com/alovajs/alova/commit/586103fefdb3ddc1ae8eab40f6c92ebab4f495a8)]:
+  - @alova/shared@1.0.0-beta.6
+
+## 3.0.0-beta.8
+
+### Patch Changes
+
+- [#430](https://github.com/alovajs/alova/pull/430) [`0929db3`](https://github.com/alovajs/alova/commit/0929db3830e656eaff90517c38da3739b0eb5917) Thanks [@JOU-amjs](https://github.com/JOU-amjs)! - fix that states updating not work with `updateStateEffect` in `alova/client`
+
+## 3.0.0-beta.7
+
+### Patch Changes
+
+- [#428](https://github.com/alovajs/alova/pull/428) [`03bca82`](https://github.com/alovajs/alova/commit/03bca821a1f0d18e5c6fc2262756f1af55108a53) Thanks [@JOU-amjs](https://github.com/JOU-amjs)! - return response data in `middleware.next` of `useSQRequest`
+
+- Updated dependencies [[`03bca82`](https://github.com/alovajs/alova/commit/03bca821a1f0d18e5c6fc2262756f1af55108a53)]:
+  - @alova/shared@1.0.0-beta.5
+
 ## 3.0.0-beta.6
 
 ### Patch Changes